getenv -> qgetenv some other win32 related fixes
svn path=/trunk/KDE/kdegames/kolf/; revision=930061

+++ game.cpp
@@ -3778,7 +3778,7 @@ void KolfGame::shotStart()
 	{ //debug code for reproducing shots, remove
 		(*curPlayer).ball()->setPos( 200, 360 );
 		strength = 6.9375;
-		float angle = -1.39094;
+		double angle = -1.39094;
 
 		(*curPlayer).ball()->collisionDetect((*curPlayer).ball()->x(), (*curPlayer).ball()->y());
